assets/requirements.txt
qrcode[pil]==8.2
python-dotenv==1.1.1
pillow==12.0.0


